requirements.txt
streamlit
pandas
google-api-python-client
transformers
torch
wordcloud
matplotlib
seaborn
pyyaml
scikit-learn
